Disabled broadcastNewFile() on each write
I can't figure out why the original developers felt it was necessary to notify the OS about the existence of the file each time it is written to.  It seems that perhaps this might be desirable once all writes are complete.  Perhaps there is no "close" or "flush" that would fulfill that purpose.  It also seems that it should be a developer's choice as to whether to notify the OS or not.  And furthermore, the rule that if a file begins with a "." or resides in a folder with ".nomedia" it should certainly be exempt from this broadcast.